- Tests fixes, - ArchivationService fixes, - GZipBlockStream fixes: corrected metod to init object and GetBlockBytes(int)

+++ src/TestTask_GZipArchiver.Core/Models/Interfaces/IBlockStream.cs
@@ -7,7 +7,7 @@ namespace TestTask_GZipArchiver.Core.Models.Interfaces
     public interface IBlockStream
     {
         int BlockSize { get; }
-        long BlocksCount { get; }
+        int BlocksCount { get; }
 
         byte[] GetBlockBytes(int blockNumber);
     }
